requirements.txt
# Python Package Index
-i https://pypi.org/simple

# Django core and utilities
django==5.2.4; python_version >= '3.10'  # Main Django framework
asgiref==3.9.1; python_version >= '3.9'  # ASGI support for Django
sqlparse==0.5.3; python_version >= '3.8'  # SQL parsing for Django
tzdata==2025.2; python_version >= '2'  # Timezone data

# Django REST Framework and extensions
djangorestframework==3.16.0; python_version >= '3.9'  # Django REST Framework
djangorestframework-simplejwt==5.5.1; python_version >= '3.9'  # JWT authentication for DRF
django-filter==25.1; python_version >= '3.9'  # Filtering for DRF
markdown==3.8.2; python_version >= '3.9'  # Markdown support for DRF browsable API

# Debugging and development tools
django-debug-toolbar==6.0.0; python_version >= '3.9'  # Debug toolbar for Django

# Image and file handling
pillow==11.3.0; python_version >= '3.9'  # Image processing

# Authentication and security
pyjwt==2.10.1; python_version >= '3.9'  # JWT token handling

# HTTP and networking
requests==2.32.4; python_version >= '3.8'  # HTTP requests
certifi==2025.7.14; python_version >= '3.7'  # SSL certificates
charset-normalizer==3.4.2; python_version >= '3.7'  # Encoding detection
idna==3.10; python_version >= '3.6'  # Internationalized domain names
urllib3==2.5.0; python_version >= '3.9'  # HTTP client
